[Model] Narrow SenseNova FP8 changes to configuration routing
Restore BF16 input image preprocessing and simplify CPU routing coverage to unquantized and online FP8 configurations.

Validation: the simplified CPU routing tests passed (2 cases); pre-commit checks passed for both changed files.

AI assistance: Codex assisted with implementation, tests, and documentation.

+++ vllm_omni/diffusion/models/sensenova_u1/pipeline_sensenova_u1.py
@@ -1048,7 +1048,7 @@ class SenseNovaU1Pipeline(
                 min_pixels=512 * 512,
                 max_pixels=max_pixels_per_image,
             )
-            pixel_values_list.append(pv.to(self.device, dtype=self.od_config.dtype))
+            pixel_values_list.append(pv.to(self.device, dtype=torch.bfloat16))
             grid_hw_list.append(ghw.to(self.device))
         return torch.cat(pixel_values_list), torch.cat(grid_hw_list)
